Move components around for cleaner structure.
